Docs maintenance (#52)
* Update ReadTheDocs build config

Upgraded the OS to Ubuntu 24.04 and updated the Python toolchain to
mambaforge-23.11 in the .readthedocs.yaml file. This ensures
compatibility with the latest environment settings available on
readthedocs.

* Update Sphinx & optional dependencies version pin

Updated Sphinx and optional dependencies packages pinned versions in
`environment-dev.yaml`, `environment-test.yaml`, and `environment-rtd.yaml` in
preparation for Python 3.13 environment.

* Upgrade blacken-docs dependency in pre-commit config

Updated the black formatter for blacken-docs from version 21.10b0 to 24.10.0.
This change ensures compatibility with the latest version of black and avoids
a version conflict problem during setup of the the pre-commit virtualenv.

* Update OceanParcels doc URL in Sphinx configuration

Change the intersphinx mapping for OceanParcels to its new documentation URL.
This ensures references to OceanParcels documentation are correctly resolved
during the build process.

* Fix xref link in Stokes_drift kernel docstring

Changed the documentation reference link in the Stokes_drift kernel to point
to the correct section in the example notebooks. This ensures users can find
more detailed usage examples without any confusion.

* Update redirected docs links found by linkcheck

Changed outdated Conda links in `docs/pkg_development.rst` and
`docs/installation.rst` to their current URLs. This ensures users are directed
to the most accurate and up-to-date documentation.

* Fix duplicated link definition in install docs

* Update docs example output from Sphinx

Updated Sphinx build and linkcheck commands example outputs.